requirements.txt
Flask==3.0.1
python-dotenv==1.0.1
twilio==8.12.0
